Last version of Sat Jul 9 @ home on F15.
Tested to add in a vtkwidget and display the cow, commemted out
the vtk components for faster build.

Removed the display panel, no need to display socket data on GUI.

Now the program can recv data and store to buffer when the user
clicks the getFortune button. The usleep() function also works
for std::cout.