I switch from qtcpsocket back to Unix socket due to recv problems. Now the client is able to work the same as the ClientConsole program.
One thing needs to solve before moving on is that the program does
not print the outputs to console in real time, instead it waits until
the program terminates and then print everything all together. This might
be caused by the signal and slot structure.

Then I can move on to incorporate the controller and add in a vtk widget.